fix(docs): use defaultEntryName instead of defaultEntryFile for Go component
defaultEntryFile matches against dist bundle paths in templateFiles,
not source file paths. Remove the incorrect defaultEntryFile prop and
rely on defaultEntryName which matches by entry name.
