+++ examples/suspense/lynx.config.ts
@@ -16,7 +16,6 @@ export default defineConfig({
   source: {
     entry: {
       main: './src/index.ts',
-      'define-async': './src/define-async.ts',
     },
   },
   plugins: [
