+++ examples/suspense/src/SlowContent.vue
@@ -1,8 +1,3 @@
-<!--
-  SlowContent.vue — Content that appears after the artificial delay.
-  The delay is applied in the defineAsyncComponent wrapper,
-  so this component itself is synchronous.
--->
 <script setup lang="ts">
 const loadedAt = new Date().toLocaleTimeString()
 </script>
